libsnapshot: use compression_factor in ota
Parse manifest compression_factor and set CowOptions appropriately. This
allows v3 writer to use compression factor in OTA. Updating some
comments about supported compression algorithms

Test: th

+++ fs_mgr/libsnapshot/include/libsnapshot/cow_writer.h
@@ -62,7 +62,7 @@ struct CowOptions {
     uint64_t op_count_max = 0;
 
     // Compression factor
-    int compression_factor = 4096;
+    uint64_t compression_factor = 4096;
 };
 
 // Interface for writing to a snapuserd COW. All operations are ordered; merges
